pyproject.toml
[project]
name = "fastapiproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.2",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "uvicorn>=0.42.0",
]
